Add support for scala 2/3 compatible macros
A lot of things were removed that the original scala-logging repo was leveraging heavily:
- Build/deploy definitions (these can easily be restored)
- LoggerTakingImplicits (couldn't get the version-unifying project to build with implicits being involved in the macros for some reason)
- Scala 2.11 and 2.12 support (cross-version macro support is limited to very specific versions of Scala 2.13 and Scala 3)

This commit represents the beginning of the Lucid Software branch of the scala-logging library. Due to the major changes, it should never be merged back into the scala-logging library, but it can serve as a starting point for those interested in attempting to bridge the compatibility gap between this library and Scala 2 + 3 in a less destructive way.